requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-RESTful==0.3.10
matplotlib==3.8.2
mutagen==1.47.0
SQLAlchemy==2.0.23
python-dotenv==1.0.0
Flask-SQLAlchemy==3.1.1